requirements.txt
boto3>=1.34.0
openai>=1.0.0
google-cloud-aiplatform>=1.38.0
streamlit>=1.28.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
pillow>=10.0.0
python-dotenv>=1.0.0
tqdm>=4.65.0
pyarrow>=12.0.0
numpy>=1.24.0
scipy>=1.11.0
